move test to integration tests
